Implement "findR" command for OmniHealth
Add FindRCommand that takes in an Index and a
RecordContainsKeywordsPredicate, and returns a CommandResult containing
the filtered record list.

Add RecordContainsKeywordsPredicate to test if a Record matches any of
the keywords given.

Add FindRCommandParser that will know how to parse multiple arguments,
one index and keywords with prefix "k/".

Modify AddressBookParser to use the newly implemented
FindRCommandParser.

Modify CliSyntax to include the keywords prefix "k/".

Add RecordList#getFilteredRecordList() method to get a RecordList that
matches a given predicate.

Add RecordList#stringifyRecordList() to get the String representation of
a RecordList.